python-microservices/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Analytics
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
faker>=20.0.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0
